+++ visualization/dot-visualizer/src/test/java/net/automatalib/visualization/dot/ProviderTest.java
@@ -57,7 +57,7 @@ public class ProviderTest {
         Assert.assertTrue(browserProvider instanceof GraphVizBrowserVisualizationProvider);
     }
 
-    // Headless GUI testing is a pain. Therefore just check that we don't throw any exceptions for now.
+    // Headless GUI testing is a pain. Therefore, just check that we don't throw any exceptions for now.
     @Test(dependsOnMethods = "testProviderConfiguration", timeOut = 10000)
     public void testDisplay() throws InterruptedException, InvocationTargetException {